Concept of FilterMixin
